cleanup/simplify local files processing
